% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{insert_dot}
\alias{insert_dot}
\title{Insert dot at the relevant place}
\usage{
insert_dot(expr, special_cases = TRUE)
}
\arguments{
\item{expr}{the quoted rhs to edit}

\item{special_cases}{wether to change \code{package::fun} to \code{package::fun()},
\code{package:::fun} to \code{package:::fun()} and \code{env$fun} to \code{env$fun()}}
}
\description{
Helper function to use in the \code{format_fun} argument of \code{new_pipe()}.
}
\details{
Note that \emph{magrittr} behaves as if \code{special_case} is \code{FALSE}. We think setting
it to \code{TRUE} is very unlikely to break any code or give an unexpected behavior
so we set it by default to \code{TRUE} and that's the way our default pipes use it.
}
