% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{is_pipe}
\alias{is_pipe}
\alias{is_compound_pipe}
\alias{is_standard_pipe}
\title{Check if object is a pipe}
\usage{
is_pipe(x)

is_compound_pipe(x)

is_standard_pipe(x)
}
\arguments{
\item{x}{An object}
}
\description{
Functions to check if an object is a pipe or a .
}
