% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipes.R
\name{pipes}
\alias{pipes}
\alias{\%>\%}
\alias{\%B>\%}
\alias{\%<>\%}
\alias{\%T>\%}
\alias{\%$\%}
\alias{\%S>\%}
\title{Pipes}
\usage{
lhs \%>\% rhs

lhs \%B>\% rhs

lhs \%<>\% rhs

lhs \%T>\% rhs

lhs \%$\% rhs

lhs \%S>\% rhs
}
\arguments{
\item{lhs}{A value or a dot (\code{.}).}

\item{rhs}{A function call using pipe semantics of the relevant pipe.}
}
\description{
Equivalents to magrittr's pipes, we also provide a \code{\%S>\%} to splice
any input and a \code{\%B>\%} pipe to debug the full pipe chain from its call.
}
